requirements.txt
scipy==1.6.2
h5py==3.2.1
numpy==1.20.2
pandas==1.2.3
torchvision==0.9.0
torch==1.8.0
Pillow==9.1.0
scikit_learn==1.0.2
